% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart-multi.R
\name{ecopart.multi}
\alias{ecopart.multi}
\title{Partitioning multi-site beta diversity}
\usage{
ecopart.multi(d1, d2, index = "whittaker", components = "four")
}
\arguments{
\item{d1}{Matrix or dataframe at time 1. Rows are sites, columns are species, and elements are presence-absence (01) or abundances of species.}

\item{d2}{Matrix or dataframe at time 2. Note that \code{d1} and \code{d2} must have exactly the same sites and species in the same order.}

\item{index}{Type of beta diversity measure. Options are \code{"whittaker"} and \code{"baselga"}.
\describe{
\item{\code{"whittaker"}}{This index is based on presence-absence data (Whittaker 1960). When \code{d1} and \code{d2} are abundance data, the elements are automatically converted to presence-absence data by replacing non-zero values with 1.}
\item{\code{"baselga"}}{This index, also known as the normalized abundance-based Whittaker's beta, is based on abundance data (Baselga 2017). When \code{d1} and \code{d2} are presence-absence data, Baselga's beta is equivalent to Harrison's beta (Harrison et al. 1992), also known as the normalized Whittaker's beta.}
}}

\item{components}{Types of components into which the total change in beta diversity is partitioned. Options are \code{"two"}, \code{"four"}, and \code{"sp"}.
\describe{
\item{\code{"two"}}{Calculates extinction and colonization components (when \code{index = "whittaker"}) or subtractive and additive components (when \code{index = "baselga"}).}
\item{\code{"four"}}{Calculates extinction- and colonization-induced homogenization and differentiation (when \code{index = "whittaker"}) or subtractive and additive homogenization and differentiation (when \code{index = "baselga"}).}
\item{\code{"sp"}}{Same with \code{"four"} but the components are further partitioned down to the species-level.}
}}
}
\value{
The \code{ecopart.multi} function returns a vector or matrix object containing the partitioned components of beta diversity.
\itemize{
\item When \code{components = "two"} was specified, the function returns a vector object with two elements: extinction and colonization components (when \code{index = "whittaker"}) or subtractive and additive components (when \code{index = "baselga"}). The extinction and colonization components represent temporal changes in beta diversity that result from local species extinctions and colonizations. The subtractive and additive components represent temporal changes in beta diversity that result from local losses and gains in species abundances.
\item When \code{components = "four"} was specified, the function returns a vector object with four elements: extinction- and colonization-induced homogenization and differentiation (when \code{index = "whittaker"}) or subtractive and additive homogenization and differentiation (when \code{index = "baselga"}). Homogenization and differentiation indicate decreases and increases in beta diversity, respectively.
\item When \code{components = "sp"} was specified, the function returns a matrix object. The rows represent the four components that are equivalent to when \code{components = "four"} was specified. The columns represent species.
}
}
\description{
\code{ecopart.multi} patitions the temporal changes in multi-site beta diversity (Whittaker or Baselga's beta) into dynamic components based on methods proposed by Tatsumi et al. (2021, 2022)
}
\references{
\itemize{
\item Baselga A (2017) Partitioning abundance-based multiple-site dissimilarity into components: balanced variation in abundance and abundance gradients. \emph{Methods in Ecology and Evolution} 8(7): 799-808.
\item Harrison S, Ross SJ, Lawton JH (1992) Beta diversity on geographic gradients in Britain. \emph{Journal of Animal Ecology} 61(1): 151-158.
\item Tatsumi S, Iritani R, Cadotte MW (2021) Temporal changes in spatial variation: partitioning the extinction and colonisation components of beta diversity. \emph{Ecology Letters} 24(5): 1063-1072.
\item Tatsumi S, Iritani R, Cadotte MW (2022) Partitioning the temporal changes in abundance-based beta diversity into loss and gain components. \emph{Methods in Ecology and Evolution}: in press
\item Whittaker RH (1960) Vegetation of the Siskiyou Mountains, Oregon and California. \emph{Ecological Monographs} 30(3): 279-338.
}
}
\author{
Shinichi Tatsumi
}
