% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart-pair.R
\name{ecopart.pair}
\alias{ecopart.pair}
\title{Partitioning pairwise dissimilarity}
\usage{
ecopart.pair(d1, d2, index = "sorensen", components = "four")
}
\arguments{
\item{d1}{A matrix or dataframe at time 1. Rows are a pair of sites (sites 1 and 2), columns are species, and elements are presence-absence (01) or abundances of species.}

\item{d2}{A matrix or dataframe at time 2. Note that \code{d1} and \code{d2} must have exactly the same sites and species in the same order.}

\item{index}{Type of dissimilarity measure. Options are \code{"jaccard"}, \code{"sorensen"}, \code{"ruzicka"}, and \code{"bray-curtis"}.
\describe{
\item{\code{"jaccard"} and \code{"sorensen"}}{These indices are based on presence-absence data (Jaccard 1912, Sorensen 1948). When \code{d1} and \code{d2} are abundance data, the elements are automatically converted to presence-absence data by replacing non-zero values with 1.}
\item{\code{"ruzicka"} and \code{"bray-curtis"}}{These indices are based on abundance data (Bray & Curtis 1957, Ruzicka 1958). When \code{d1} and \code{d2} are presence-absence data, Ruzicka and Bray-Curtis indices are equivalent to the Jaccard and Sorensen indices, respectively.}
}}

\item{components}{Types of components into which the total change in beta diversity is partitioned. Options are \code{"two"}, \code{"four"}, \code{"six"}, and \code{"sp"}.
\describe{
\item{\code{"two"}}{Calculates extinction and colonization components (when \code{index = "jaccard"} or \code{index = "sorensen"}) or subtractive and additive components (when \code{"ruzicka"} or \code{"bray-curtis"}).}
\item{\code{"four"}}{Calculates extinction- and colonization-induced homogenization and differentiation (when \code{index = "jaccard"} or \code{index = "sorensen"}) or subtractive and additive homogenization and differentiation (when \code{"ruzicka"} or \code{"bray-curtis"}).}
\item{\code{"six"}}{Calculates six components that represent (1) homogenization by decrease in U, (2) differentiation by decrease in C, (3) differentiation by change from C to U, (4) differentiation by increase in U, (5) homogenization by increase in C, and (6) homogenization by change from U to C, where U is the existance or abundance of species unique to either site and C is the existance or abundance of species common to both sites.}
\item{\code{"sp"}}{Same with \code{"four"} but the components are further partitioned down to the species-level.}
}}
}
\value{
The \code{ecopart.pair} function returns a vector or matrix object containing the partitioned components of beta diversity.
\itemize{
\item When \code{components = "two"} was specified, the function returns a vector object with two elements: extinction and colonization components (when \code{index = "jaccard"} or \code{index = "sorensen"}) or subtractive and additive components (when \code{index = "ruzicka"} or \code{index = "bray-curtis"}). The extinction and colonization components represent temporal changes in beta diversity that result from local species extinctions and colonizations. The subtractive and additive components represent temporal changes in beta diversity that result from local losses and gains in species abundances.
\item When \code{components = "four"} was specified, the function returns a vector object with four elements: extinction- and colonization-induced homogenization and differentiation (when \code{index = "jaccard"} or \code{index = "sorensen"}) or subtractive and additive homogenization and differentiation (when \code{index = "ruzicka"} or \code{index = "bray-curtis"}). Homogenization and differentiation indicate decreases and increases in beta diversity, respectively.
\item When \code{components = "six"} was specified, the function returns a vector object with six elements: (1) homogenization by decrease in U, (2) differentiation by decrease in C, (3) differentiation by change from C to U, (4) differentiation by increase in U, (5) homogenization by increase in C, and (6) homogenization by change from U to C.
\item When \code{components = "sp"} was specified, the function returns a matrix object. The rows represent the four components that are equivalent to when \code{components = "four"} was specified. The columns represent species.
}
}
\description{
\code{ecopart.pair} patitions the temporal changes in pairwise dissimilarity (Jaccard, Sorensen, Ruzicka, and Bray-Curtis indices) into dynamic components based on methods proposed by Tatsumi et al. (2021, 2022)
}
\references{
\itemize{
\item Bray JR, Curtis JT (1957) An ordination of the upland forest communities of southern Wisconsin. \emph{Ecological Monographs} 27(4): 325-349.
\item Jaccard P (1912) The distribution of the flora in the alphine zone. \emph{New Phytologist} 11(2): 37-50.
\item Ruzicka M (1958) Anwendung mathematisch-statisticher Methoden in der Geobotanik (synthetische Bearbeitung von Aufnahmen). \emph{Biologia, Bratislava} 13: 647-661.
\item Sorensen T (1948) A method of establishing groups of equal amplitude in plant sociology based on similarity of species and its application to analyses of the vegetation on Danish commons. \emph{Kongelige Danske Videnskabernes Selskab} 5(4): 1-34.
\item Tatsumi S, Iritani R, Cadotte MW (2021) Temporal changes in spatial variation: partitioning the extinction and colonisation components of beta diversity. \emph{Ecology Letters} 24(5): 1063-1072.
\item Tatsumi S, Iritani R, Cadotte MW (2022) Partitioning the temporal changes in abundance-based beta diversity into loss and gain components. \emph{Methods in Ecology and Evolution}: in press
}
}
\author{
Shinichi Tatsumi
}
