% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopart.R
\name{ecopart.pair}
\alias{ecopart.pair}
\title{Partition beta diversity}
\usage{
ecopart.pair(d1, d2, index = "ruzicka", components = "two")
}
\arguments{
\item{d1}{a}

\item{d2}{b}

\item{index}{c}

\item{components}{d}
}
\description{
Partition beta diversity
}
